+++ WooCommerce/Classes/ViewRelated/Orders/Order Details/Customer Section/CustomerNoteTableViewCell.swift
@@ -130,7 +130,7 @@ private extension CustomerNoteTableViewCell {
         addButton.contentVerticalAlignment = .bottom
         var configuration = UIButton.Configuration.plain()
         configuration.contentInsets = .init(.zero)
-        configuration.imagePadding = CGFloat(Constants.buttonTitleAndImageSpacing)
+        configuration.imagePadding = Constants.buttonTitleAndImageSpacing
         addButton.configuration = configuration
         addButton.addTarget(self, action: #selector(addButtonTapped), for: .touchUpInside)
     }
